fix more errors

+++ src-scala/org/jgll/util/hashing/IntegerExternalHasher.scala
@@ -1,10 +1,7 @@
 package org.jgll.util.hashing
 
 import org.jgll.util.hashing.hashfunction.HashFunction
-import IntegerExternalHasher._
-import scala.reflect.{BeanProperty, BooleanBeanProperty}
-//remove if not needed
-import scala.collection.JavaConversions._
+import scala.reflect.BeanProperty
 
 object IntegerExternalHasher {
 
@@ -13,9 +10,9 @@ object IntegerExternalHasher {
 }
 
 @SerialVersionUID(1L)
-class IntegerExternalHasher private () extends ExternalHasher[Integer] {
+class IntegerExternalHasher private () extends ExternalHasher[Int] {
 
-  override def hash(t: java.lang.Integer, f: HashFunction): Int = f.hash(t)
+  override def hash(t: Int, f: HashFunction): Int = f.hash(t)
 
-  override def equals(t1: java.lang.Integer, t2: java.lang.Integer): Boolean = t1.intValue() == t2.intValue()
+  override def equals(t1: Int, t2: Int): Boolean = t1 == t2
 }
